more

+++ lang/cim/c11.hpp
@@ -7,7 +7,7 @@ void c11yy_iconst_readsuffix(struct c11yy_struct_integer &val,const char* &s);
 void c11yy_iconst_read(const unsigned int base,struct c11yy_struct_integer &val,const char* &s);
 void c11yy_iconst_readchar(const enum c11yystringtype st,struct c11yy_struct_integer &val,const char* &s);
 void c11yy_check_stringtype_prefix(enum c11yystringtype &st,const char* &s);
-int64_t c11yy_read_utf8(const char* &s);
+uint32_t c11yy_read_utf8(const char* &s);
 
 #endif
